stop-script

+++ scripts/stop_container.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 set -e
 
-# Stop the running container (if any)
 containerid=$(docker ps -q)
-docker rm -f $containerid
+
+if [ ! -z "$containerid" ]; then
+    docker rm -f $containerid
+fi
